refactor(app): 拆分主入口壳层与对话管理页
