+++ pinvou3-app/src/app/main.jsx
@@ -2,8 +2,8 @@ import React, { useCallback, useEffect, useLayoutEffect, useRef, useState } from
 import { createPortal } from 'react-dom';
 import { createRoot } from 'react-dom/client';
 import '../styles/base.css';
-import { I, Plus, Edit2, Trash2, ClipboardList, BarChart2, Settings, Monitor, Smartphone, Brain, BrainCircuit, Clock, Sun, Moon, Zap, Package, RefreshCw, RotateCcw, Search, Upload, Lightbulb, Paperclip, Mic, Send, Store, Terminal, ChevronDown, IconGrid, IconList, Copy, CheckCircle2, AlertTriangle, Menu, MoreHorizontal, Check, Filter, Database, Download, FolderPlus, Award, Feather, AppWindow, Radio, Palette, Briefcase, StopCircle, XCircle, Wrench, Layers, MessageSquare, X, ArrowLeft, FolderOpen, ExternalLink, BookOpen, Code, FileText, Hexagon, Layout, Presentation, Mail, MessageCircle, Navigation, Video, Puzzle, LineChart, Building2, Cpu, Server, Globe, ChevronLeft, XIcon, CloudSun, TrendingUp, TrendingDown, GridIcon, TableIcon, PresentationIcon, ImageIcon, Archive, PinIcon, PinOffIcon } from '../components/icons.jsx';
-import { ArchiveConfirmDialog, ArchiveToast, ArchivedDeleteConfirmDialog, NavItem, RecentItem } from '../components/layout/NavigationComponents.jsx';
+import { I, Plus, Edit2, Trash2, ClipboardList, BarChart2, Settings, Monitor, Smartphone, Brain, BrainCircuit, Clock, Sun, Moon, Zap, Package, RefreshCw, RotateCcw, Search, Upload, Lightbulb, Paperclip, Mic, Send, Store, Terminal, ChevronDown, IconGrid, IconList, Copy, CheckCircle2, AlertTriangle, Menu, MoreHorizontal, Check, Filter, Database, Download, FolderPlus, Award, Feather, AppWindow, Radio, Palette, Briefcase, StopCircle, XCircle, Wrench, Layers, MessageSquare, X, ArrowLeft, FolderOpen, ExternalLink, BookOpen, Code, FileText, Hexagon, Layout, Presentation, Mail, MessageCircle, Navigation, Video, Puzzle, LineChart, Building2, Cpu, Server, Globe, ChevronLeft, XIcon, CloudSun, TrendingUp, TrendingDown, GridIcon, TableIcon, PresentationIcon, ImageIcon, Archive, PinIcon, PinOffIcon, PetPawIcon } from '../components/icons.jsx';
+import { ArchiveConfirmDialog, ArchiveToast, NavItem, RecentItem } from '../components/layout/NavigationComponents.jsx';
 import { AcpAgentLogo } from '../features/codex/AcpAgentLogo.jsx';
 import { PinvouLogo } from '../components/PinvouLogo.jsx';
 import { MobileMoreSheet, MobileTabBar, MobileTopBar } from '../components/layout/MobileShell.jsx';
@@ -12,7 +12,7 @@ import { bridge, useBridgeState, activeModelIsLocal, shouldShowApiKeyGate } from
 import { useCompactViewport, useVisualViewportHeight } from '../hooks/useViewport.js';
 import { dict, LANG_TO_TAG, SEARCH_KEY_PROVIDERS, TAG_TO_LANG } from '../shared/i18n.js';
 import { formatSessionDate, localDateKey, formatDateGroupLabel } from '../shared/date-utils.js';
-import { runSessionBatch, sessionRoute } from '../shared/session-management.js';
+import { runSessionBatch } from '../shared/session-management.js';
 import { can, isWeb } from '../shared/platform.js';
 import { installGlobalMarkdownRenderer } from '../shared/markdown-renderer.js';
 import { KnowledgeView } from '../features/knowledge/KnowledgeView.jsx';
@@ -34,7 +34,6 @@ import {
   isTauriAvailable,
   tauriCommands,
   tauriEvents,
-  tryGetCurrentTauriWindow,
 } from '../platform/tauri/client.js';
 
 // 定时任务创建与运行链路已恢复，展示入口并允许自动跳转。
@@ -53,7 +52,9 @@ import { ToolStoreView } from '../features/tools/ToolStoreView.jsx';
 import { PinvouSummonCard } from '../features/tools/tool-renderers.jsx';
 import { CardPoolView, Lanyard, PersonaEditorModal } from '../features/personas/Personas.jsx';
 import { WorkflowView } from '../features/workflow/WorkflowView.jsx';
+import { SearchView } from '../features/search/SearchView.jsx';
 import { DetachedShell } from './DetachedShell.jsx';
+import { TitleBar } from './DesktopTitleBar.jsx';
 
 installGlobalMarkdownRenderer(window);
 window.__PINVOU_STARTUP__.mark('app:main_module_body_enter');
@@ -87,71 +88,6 @@ function workspaceDisplayName(path) {
   const parts = String(path || '').split(/[\\/]/).filter(Boolean);
   return parts[parts.length - 1] || String(path || '');
 }
-
-/* ==========================================
-       Lucide icon replacements (inline SVG)
-       ========================================== */
-    const appWindow = tryGetCurrentTauriWindow();
-    const TitleBar = ({ theme, t, sidebarOpen = true }) => {
-      const isDark = theme === 'dark';
-      const hoverBg = isDark ? 'hover:bg-white/10' : 'hover:bg-black/10';
-      const titleBarBg = isDark ? (sidebarOpen ? 'bg-[#1E1F20]' : 'bg-[#131314]') : 'bg-[#F0F4F9]';
-      // macOS 顶栏走系统原生实现:窗口带 decorations + titleBarStyle=Overlay
-      // (见 src-tauri/config/platforms/macos/tauri.conf.json),系统红绿灯悬浮在内容区左上角,
-      // 此时不再渲染 Windows 风格三键,并为红绿灯留出左侧空间。
-      // 红绿灯纵向位置由 overlay 配置 trafficLightPosition y=20 决定:tao 按
-      // "容器高 = 按钮高(14) + y" 布局,按钮顶边距窗口顶 y-9,按钮圆心 7,
-      // 故 y=20 时圆心 18,正好与本 h-9(36px)顶栏内容中线对齐;调整顶栏高度需同步改 y。
-      // Windows/Linux 窗口无边框(decorations=false),继续用自绘三键。
-      // 以窗口实际 decorations 状态为准,不解析 UA / 平台字符串。
-      // 初始 null 表示探测未决:此时不渲染自绘三键,避免 macOS 原生顶栏下
-      // 三键先闪现一帧再被隐藏;探测失败回退为自绘三键(fail-safe)。
-      const [nativeControls, setNativeControls] = useState(null);
-      useEffect(() => {
-        let cancelled = false;
-        if (appWindow && typeof appWindow.isDecorated === 'function') {
-          appWindow.isDecorated()
-            .then((decorated) => { if (!cancelled) setNativeControls(decorated === true); })
-            .catch(() => { if (!cancelled) setNativeControls(false); });
-        } else {
-          setNativeControls(false);
-        }
-        return () => { cancelled = true; };
-      }, []);
-      return (
-        <div data-tauri-drag-region
-          className={`h-9 shrink-0 flex items-center justify-between select-none ${titleBarBg} ${isDark ? 'text-[#E3E3E3]' : 'text-[#1F1F1F]'}`}>
-          <div data-tauri-drag-region className={`flex items-center gap-2 ${nativeControls === true ? 'pl-[76px] pr-3' : 'px-3'} text-[13px] font-medium pointer-events-none`}>
-            <PinvouLogo className="h-[18px] w-[18px] select-none" />
-            {t.appTitle}
-          </div>
-          {nativeControls === false && (
-          <div className="flex items-center h-full">
-            <button onClick={() => appWindow && appWindow.minimize()} title={t.winMin}
-              className={`h-full w-12 flex items-center justify-center transition-colors ${hoverBg}`}>
-              <svg width="11" height="11" viewBox="0 0 11 11"><rect x="1" y="5" width="9" height="1" fill="currentColor"/></svg>
-            </button>
-            <button onClick={() => appWindow && appWindow.toggleMaximize()} title={t.winMax}
-              className={`h-full w-12 flex items-center justify-center transition-colors ${hoverBg}`}>
-              <svg width="11" height="11" viewBox="0 0 11 11"><rect x="1.5" y="1.5" width="8" height="8" fill="none" stroke="currentColor" strokeWidth="1"/></svg>
-            </button>
-            <button onClick={() => appWindow && appWindow.close()} title={t.winClose}
-              className="h-full w-12 flex items-center justify-center transition-colors hover:bg-[#E81123] hover:text-white">
-              <svg width="11" height="11" viewBox="0 0 11 11"><path d="M1 1 L10 10 M10 1 L1 10" stroke="currentColor" strokeWidth="1.1"/></svg>
-            </button>
-          </div>
-          )}
-        </div>
-      );
-    };
-
-    // 桌宠快捷开关的爪印图标(lucide paw-print 风格,icons.jsx 没有现成的)
-    const PetPawIcon = () => (
-      <svg width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-        <circle cx="11" cy="4" r="2" /><circle cx="18" cy="8" r="2" /><circle cx="4" cy="8" r="2" />
-        <path d="M14.35 13.5a4 4 0 0 0-6.7 0c-.63 1.4-1.6 2.4-2.25 3.3-.72 1 .16 3.2 1.9 3.2 1.4 0 2.3-1 4.35-1s2.95 1 4.35 1c1.74 0 2.62-2.2 1.9-3.2-.65-.9-1.62-1.9-2.25-3.3z" />
-      </svg>
-    );
 
     /* ==========================================
        MegaCube 本地大模型引导:进行中步骤指示 + 自跑计时
@@ -2143,9 +2079,6 @@ function workspaceDisplayName(path) {
             {currentView === 'search' && (
               <SearchView
                 theme={activeTheme} history={allSidebarTasks} t={t} language={language}
-                activeId={activeChat}
-                activeCodexId={activeCodexId}
-                activeScheduledId={(bs && bs.scheduledRunContext && bs.scheduledRunContext.sessionId) || ''}
                 archived={(bs && bs.archivedSessions) || []}
                 showArchived={searchShowArchived}
                 onShowArchivedConsumed={() => setSearchShowArchived(false)}
@@ -2587,517 +2520,6 @@ function workspaceDisplayName(path) {
       );
     };
 
-    // 对话管理页:上方搜索框,工具行「对话|已收纳」切换 + 批量管理,左侧日期栏,右侧对话列表。
-    // 已收纳与在线对话共用同一套日期分组/搜索/多选管线,仅数据源与行操作不同
-    // (在线:置顶/重命名/收纳/删除;已收纳:恢复/永久删除,按对话自身更新时间分组)。
-    // 无搜索词:右侧显示所选日期的对话;有搜索词:右侧按日期分组显示全部匹配项,
-    // 左侧只保留有匹配的日期,点击日期平滑滚动到右侧对应分组。
-    const SearchView = ({ theme, history, t, language, activeId, activeCodexId, activeScheduledId, archived = [], showArchived: showArchivedProp, onShowArchivedConsumed, onSelect, onOpenCodex, onOpenScheduledRun, onRename, onDelete, onTogglePinned, onOpenFolder, onArchive, onArchiveMany, onDeleteMany, onRestoreArchived, onRestoreMany }) => {
-      const isDark = theme === 'dark';
-      const [query, setQuery] = useState('');
-      const [selectedDate, setSelectedDate] = useState(null);
-      const [showArchived, setShowArchived] = useState(false);
-      const [batchMode, setBatchMode] = useState(false);
-      const [selectedIds, setSelectedIds] = useState(() => new Set());
-      const [batchDeleteConfirming, setBatchDeleteConfirming] = useState(false);
-      const [archivedDeleteConfirm, setArchivedDeleteConfirm] = useState(null);
-      const [listFilter, setListFilter] = useState('all'); // all | pinned | scheduled(仅对话面板生效,与侧栏任务列表同款)
-      const [listSort, setListSort] = useState('pinned_first'); // pinned_first | recent
-      const [filterOpen, setFilterOpen] = useState(false);
-      const filterRef = useRef(null);
-      const inputRef = useRef(null);
-      const listRef = useRef(null);
-      const groupRefs = useRef({});
-
-      // 收纳 toast「前往查看」的一次性信号:展开「已收纳」面板后立刻消费,避免下次进页又自动打开
-      useEffect(() => {
-        if (showArchivedProp) {
-          setShowArchived(true);
-          onShowArchivedConsumed && onShowArchivedConsumed();
-        }
-      }, [showArchivedProp]);
-
-      const archivedList = archived || [];
-      // 已收纳复用同一套日期分组管线:归一成 history 形状(updatedAt 取对话自身更新时间),
-      // 原始 DTO 挂 raw 供恢复/删除按钮取 archived_at 等信息
-      const archivedHistory = archivedList.map(s => {
-        const titlePresentation = sessionTitlePresentation(s.title || t.newChat, s.title_attachment_names);
-        return {
-          id: s.id,
-          title: sessionTitlePlainText(titlePresentation),
-          titleContent: titlePresentation.attachments.length
-            ? <SessionAttachmentTitle presentation={titlePresentation} />
-            : null,
-          date: formatSessionDate(s.updated_at || s.created_at, language),
-          updatedAt: s.updated_at || s.created_at || '',
-          raw: s,
-        };
-      });
-      // 筛选/排序与左侧任务列表同款(仅对话面板):全部/置顶/定时任务 + 置顶优先/最近更新
-      const searchFilterOptions = [
-        { id: 'all', label: t.sidebarTaskFilterAll },
-        { id: 'pinned', label: t.sidebarTaskFilterPinned },
-        { id: 'scheduled', label: t.sidebarTaskFilterScheduled },
-      ];
-      const searchSortOptions = [
-        { id: 'pinned_first', label: t.sidebarTaskSortPinnedFirst },
-        { id: 'recent', label: t.sidebarTaskSortRecent },
-      ];
-      const sourceHistory = (showArchived ? archivedHistory : (history || []))
-        .filter(c => {
-          if (showArchived || listFilter === 'all') return true;
-          if (listFilter === 'pinned') return !!c.pinned;
-          if (listFilter === 'scheduled') return c.taskKind === 'scheduled';
-          return true;
-        })
-        .sort((a, b) => {
-          if (showArchived || listSort === 'recent') {
-            return String(b.updatedAt || b.pinnedAt || '').localeCompare(String(a.updatedAt || a.pinnedAt || ''));
-          }
-          if (!!a.pinned !== !!b.pinned) return a.pinned ? -1 : 1;
-          const aTime = a.pinned ? (a.pinnedAt || a.updatedAt) : (a.updatedAt || a.pinnedAt);
-          const bTime = b.pinned ? (b.pinnedAt || b.updatedAt) : (b.updatedAt || b.pinnedAt);
-          return String(bTime || '').localeCompare(String(aTime || ''));
-        });
-
-      // 筛选菜单:点外部/Escape 关闭(与侧栏筛选菜单同款交互)
-      useEffect(() => {
-        if (!filterOpen) return undefined;
-        const closeOnPointerDown = (event) => {
-          if (filterRef.current && !filterRef.current.contains(event.target)) setFilterOpen(false);
-        };
-        const closeOnEscape = (event) => {
-          if (event.key === 'Escape') {
-            event.preventDefault();
-            setFilterOpen(false);
-          }
-        };
-        document.addEventListener('pointerdown', closeOnPointerDown);
-        window.addEventListener('keydown', closeOnEscape);
-        return () => {
-          document.removeEventListener('pointerdown', closeOnPointerDown);
-          window.removeEventListener('keydown', closeOnEscape);
-        };
-      }, [filterOpen]);
-
-      // 按本地日历日分组:组内时间倒序,组间日期倒序,无时间戳落 'unknown' 沉底
-      const dateGroups = [];
-      {
-        const byDate = new Map();
-        // sourceHistory 已按当前面板排好序(置顶优先/最近更新),组内顺序即排序结果
-        sourceHistory.forEach(chat => {
-          const key = localDateKey(chat.updatedAt);
-          if (!byDate.has(key)) byDate.set(key, []);
-          byDate.get(key).push(chat);
-        });
-        byDate.forEach((rows, key) => dateGroups.push({ key, rows }));
-        dateGroups.sort((a, b) => {
-          if (a.key === 'unknown') return 1;
-          if (b.key === 'unknown') return -1;
-          return b.key.localeCompare(a.key);
-        });
-      }
-      // 'all' = 全部日期;未选或所选日期已不存在时,默认落在最近一天
-      const activeDate = selectedDate === 'all'
-        ? 'all'
-        : dateGroups.some(g => g.key === selectedDate)
-          ? selectedDate
-          : (dateGroups[0] ? dateGroups[0].key : null);
-
-      const q = query.trim().toLowerCase();
-      const searching = q.length > 0;
-      const matchedGroups = searching
-        ? dateGroups
-            .map(g => ({ key: g.key, rows: g.rows.filter(c => c.title.toLowerCase().includes(q)) }))
-            .filter(g => g.rows.length > 0)
-        : dateGroups;
-      // 左侧日期栏:搜索时只保留有匹配项的日期
-      const railGroups = searching ? matchedGroups : dateGroups;
-      const railTotal = railGroups.reduce((n, g) => n + g.rows.length, 0);
-      const activeGroup = dateGroups.find(g => g.key === activeDate) || null;
-
-      // 当前右侧可见的会话 id(全选范围):搜索态=全部匹配,否则=所选日期(或「全部」)的分组
-      const visibleGroups = searching ? matchedGroups : (activeDate === 'all' ? dateGroups : (activeGroup ? [activeGroup] : []));
-      const visibleIds = visibleGroups.flatMap(g => g.rows.map(c => c.id));
-      const allVisibleSelected = visibleIds.length > 0 && visibleIds.every(id => selectedIds.has(id));
-
-      const exitBatch = () => { setBatchMode(false); setSelectedIds(new Set()); setBatchDeleteConfirming(false); };
-      const toggleSelect = (id) => setSelectedIds(prev => {
-        const next = new Set(prev);
-        if (next.has(id)) next.delete(id); else next.add(id);
-        return next;
-      });
-      const toggleSelectAll = () => setSelectedIds(allVisibleSelected ? new Set() : new Set(visibleIds));
-      const switchPanel = (toArchived) => { setShowArchived(toArchived); exitBatch(); };
-      const runBatchArchive = () => { if (onArchiveMany) onArchiveMany(Array.from(selectedIds)); exitBatch(); };
-      const runBatchDelete = () => { if (onDeleteMany) onDeleteMany(Array.from(selectedIds)); exitBatch(); };
-      const runBatchRestore = () => { if (onRestoreMany) onRestoreMany(Array.from(selectedIds)); exitBatch(); };
-
-      const handlePickDate = (key) => {
-        setSelectedDate(key);
-        if (!searching) return;
-        // 搜索态:右侧是全部匹配的长列表,点击日期滚动到对应分组,「全部」回顶部
-        const container = listRef.current;
-        if (!container) return;
-        if (key === 'all') { container.scrollTo({ top: 0, behavior: 'smooth' }); return; }
-        const el = groupRefs.current[key];
-        if (el) container.scrollTo({ top: el.offsetTop, behavior: 'smooth' });
-      };
-
-      const renderChatRow = (chat) => {
-        if (batchMode) {
-          const selected = selectedIds.has(chat.id);
-          return (
-            <div
-              key={chat.id}
-              onClick={() => toggleSelect(chat.id)}
-              className={`flex items-center gap-3 px-4 py-[10px] cursor-pointer rounded-[16px] transition-colors ${isDark ? 'hover:bg-[#1E1F20]' : 'hover:bg-[#F0F4F9]'}`}
-            >
-              <span className={`w-5 h-5 shrink-0 rounded-full border flex items-center justify-center transition-colors ${
-                selected
-                  ? 'bg-[#0B57D0] border-[#0B57D0] text-white'
-                  : (isDark ? 'border-[#5F6368]' : 'border-[#C4C7C5]')
-              }`}>
-                {selected && <Check size={13} />}
-              </span>
-              {chat.pinned && <PinIcon size={12} className={`shrink-0 rotate-45 ${isDark ? 'text-[#9AA0A6]' : 'text-[#8A8F94]'}`} />}
-              <span className={`flex-1 min-w-0 truncate text-[15px] ${isDark ? 'text-[#E3E3E3]' : 'text-[#1F1F1F]'}`}>{chat.titleContent || chat.title}</span>
-              <span className={`text-[13px] shrink-0 ${isDark ? 'text-[#C4C7C5]' : 'text-[#444746]'}`}>{chat.date}</span>
-            </div>
-          );
-        }
-        if (showArchived) {
-          // 已收纳行:恢复 / 永久删除(删除走 ArchivedDeleteConfirmDialog 二次确认)
-          const s = chat.raw || chat;
-          return (
-            <div
-              key={chat.id}
-              className={`flex items-center gap-2 px-4 py-[12px] rounded-[16px] transition-colors ${isDark ? 'hover:bg-[#1E1F20]' : 'hover:bg-[#F0F4F9]'}`}
-            >
-              <span className="flex-1 min-w-0 pr-2">
-                <span className={`block truncate text-[15px] ${isDark ? 'text-[#E3E3E3]' : 'text-[#1F1F1F]'}`}>{chat.titleContent || chat.title}</span>
-                <span className={`block truncate text-[12px] ${isDark ? 'text-[#9AA0A6]' : 'text-[#8A8F94]'}`}>
-                  {t.searchArchivedAt(formatSessionDate(s.archived_at || s.updated_at || s.created_at, language))}
-                </span>
-              </span>
-              <button
-                type="button"
-                onClick={() => onRestoreArchived && onRestoreArchived(chat.id)}
-                className={`shrink-0 h-8 px-3 rounded-full text-[13px] font-medium transition-colors ${isDark ? 'bg-[#A8C7FA] text-[#041E49]' : 'bg-[#D3E3FD] text-[#041E49]'}`}
-              >
-                {t.searchRestore}
-              </button>
-              <button
-                type="button"
-                onClick={() => setArchivedDeleteConfirm(s)}
-                className={`shrink-0 h-8 px-3 rounded-full text-[13px] font-medium transition-colors ${isDark ? 'text-[#F28B82] hover:bg-[#5c2b29]' : 'text-[#C5221F] hover:bg-[#FAD2CF]'}`}
-              >
-                {t.cpDelete}
-              </button>
-            </div>
-          );
-        }
-        // 与左侧任务列表同款行项目:置顶/重命名/删除/收纳/打开文件夹/右键菜单;
-        // 定时任务运行项点按进入运行会话(与侧栏一致)。
-        // 管理页只是会话清单,不高亮当前会话(active 恒 false),避免打开过的对话残留"选中"背景。
-        const route = sessionRoute(chat);
-        return (
-          <RecentItem
-            key={chat.id}
-            chat={chat}
-            theme={theme}
-            t={t}
-            active={false}
-            personaTarget={false}
-            onSelect={route === 'codex'
-              ? () => onOpenCodex && onOpenCodex(chat.id)
-              : route === 'scheduled'
-                ? () => onOpenScheduledRun && onOpenScheduledRun(chat.scheduledRun)
-                : onSelect}
-            onRename={onRename}
-            onDelete={onDelete}
-            onTogglePinned={onTogglePinned}
-            onOpenFolder={onOpenFolder}
-            onArchive={onArchive}
-          />
-        );
-      };
-
-      const renderDateHeader = (key) => (
-        <div className={`px-4 pt-4 pb-1 text-[13px] font-medium ${isDark ? 'text-[#9AA0A6]' : 'text-[#8A8F94]'}`}>
-          {formatDateGroupLabel(key, language)}
-        </div>
-      );
-
-      return (
-        <div className="flex-1 flex flex-col w-full h-full relative z-10 animate-in fade-in duration-300">
-          <div className="flex-1 min-h-0 flex flex-col px-6 pt-16 pb-6">
-            <div className="max-w-[960px] w-full mx-auto flex flex-col flex-1 min-h-0 relative">
-
-              {/* Centered Search Bar */}
-              <div className={`shrink-0 flex items-center gap-3 px-6 py-4 rounded-full mb-4 transition-colors ${isDark ? 'bg-[#1E1F20] text-[#E3E3E3]' : 'bg-[#F0F4F9] text-[#1F1F1F]'}`}>
-                <Search size={22} className={isDark ? 'text-[#C4C7C5]' : 'text-[#444746]'} />
-                <input
-                  ref={inputRef}
-                  type="text"
-                  placeholder={t.searchPlaceholder}
-                  value={query}
-                  onChange={e => setQuery(e.target.value)}
-                  className={`flex-1 bg-transparent border-none outline-none text-[16px] placeholder:text-[16px] ${isDark ? 'placeholder:text-[#C4C7C5]' : 'placeholder:text-[#444746]'}`}
-                />
-                {query ? (
-                  <button
-                    type="button"
-                    aria-label={t.clearSearch}
-                    title={t.clearSearch}
-                    onClick={() => { setQuery(''); inputRef.current && inputRef.current.focus(); }}
-                    className={`w-8 h-8 shrink-0 rounded-full flex items-center justify-center transition-colors ${isDark ? 'text-[#C4C7C5] hover:bg-[#333537]' : 'text-[#444746] hover:bg-[#DDE3EA]'}`}
-                  >
-                    <X size={16} />
-                  </button>
-                ) : null}
-              </div>
-
-              {/* 工具行:「对话|已收纳」切换(批量模式下换成全选) + 批量管理开关;与搜索框同宽对齐 */}
-              <div className="shrink-0 h-9 mb-2 flex items-center justify-between">
-                {batchMode ? (
-                  <button
-                    type="button"
-                    onClick={toggleSelectAll}
-                    className={`h-8 px-2 flex items-center gap-2 rounded-full text-[13px] transition-colors ${isDark ? 'text-[#C4C7C5] hover:bg-[#1E1F20]' : 'text-[#444746] hover:bg-[#F0F4F9]'}`}
-                  >
-                    <span className={`w-4 h-4 shrink-0 rounded-full border flex items-center justify-center transition-colors ${
-                      allVisibleSelected
-                        ? 'bg-[#0B57D0] border-[#0B57D0] text-white'
-                        : (isDark ? 'border-[#5F6368]' : 'border-[#C4C7C5]')
-                    }`}>
-                      {allVisibleSelected && <Check size={11} />}
-                    </span>
-                    <span>{t.searchSelectAll} · {t.searchSelectedCount(selectedIds.size)}</span>
-                  </button>
-                ) : (
-                  <div className={`flex items-center gap-0.5 p-0.5 rounded-full ${isDark ? 'bg-[#1E1F20]' : 'bg-[#F0F4F9]'}`}>
-                    {[{ key: false, label: t.searchPanelChats }, { key: true, label: `${t.searchArchivedEntry} (${archivedList.length})` }].map(tab => (
-                      <button
-                        key={String(tab.key)}
-                        type="button"
-                        onClick={() => switchPanel(tab.key)}
-                        className={`h-7 px-3 rounded-full text-[13px] font-medium transition-colors ${
-                          showArchived === tab.key
-                            ? (isDark ? 'bg-[#A8C7FA] text-[#041E49]' : 'bg-white text-[#1F1F1F] shadow-sm')
-                            : (isDark ? 'text-[#C4C7C5]' : 'text-[#444746]')
-                        }`}
-                      >
-                        {tab.label}
-                      </button>
-                    ))}
-                  </div>
-                )}
-                <div className="flex items-center gap-1">
-                  {!showArchived && (
-                    <div ref={filterRef} className="relative">
-                      <button
-                        type="button"
-                        title={t.sidebarTaskFilter}
-                        onClick={() => setFilterOpen(v => !v)}
-                        className={`w-8 h-8 shrink-0 rounded-full flex items-center justify-center transition-colors ${
-                          filterOpen || listFilter !== 'all' || listSort !== 'pinned_first'
-                            ? (isDark ? 'bg-[#333537] text-[#E3E3E3]' : 'bg-[#E1E5EA] text-[#444746]')
-                            : (isDark ? 'text-[#C4C7C5] hover:bg-[#1E1F20]' : 'text-[#444746] hover:bg-[#F0F4F9]')
-                        }`}
-                      >
-                        <Filter size={15} />
-                      </button>
-                      {filterOpen && (
-                        <div className={`absolute right-0 top-9 z-50 w-44 overflow-hidden rounded-2xl border p-1.5 shadow-xl ${isDark ? 'border-white/10 bg-[#202124]' : 'border-black/10 bg-white'}`}>
-                          <div className={`px-2.5 pb-1 pt-1 text-[11px] font-semibold ${isDark ? 'text-[#8E8E93]' : 'text-[#8A8A8E]'}`}>
-                            {t.sidebarTaskFilter}
-                          </div>
-                          {searchFilterOptions.map(option => (
-                            <button
-                              key={option.id}
-                              type="button"
-                              onClick={() => setListFilter(option.id)}
-                              className={`w-full px-2.5 py-1.5 flex items-center gap-2 rounded-xl text-left text-[13px] leading-5 transition-colors ${isDark ? 'text-[#E3E3E3] hover:bg-[#303134]' : 'text-[#1F1F1F] hover:bg-[#F1F3F4]'}`}
-                            >
-                              <span className="w-4 shrink-0">{listFilter === option.id && <Check size={13} />}</span>
-                              <span className="truncate">{option.label}</span>
-                            </button>
-                          ))}
-                          <div className={`my-1 h-px ${isDark ? 'bg-white/10' : 'bg-black/10'}`} />
-                          <div className={`px-2.5 pb-1 pt-1 text-[11px] font-semibold ${isDark ? 'text-[#8E8E93]' : 'text-[#8A8A8E]'}`}>
-                            {t.sidebarTaskSort}
-                          </div>
-                          {searchSortOptions.map(option => (
-                            <button
-                              key={option.id}
-                              type="button"
-                              onClick={() => setListSort(option.id)}
-                              className={`w-full px-2.5 py-1.5 flex items-center gap-2 rounded-xl text-left text-[13px] leading-5 transition-colors ${isDark ? 'text-[#E3E3E3] hover:bg-[#303134]' : 'text-[#1F1F1F] hover:bg-[#F1F3F4]'}`}
-                            >
-                              <span className="w-4 shrink-0">{listSort === option.id && <Check size={13} />}</span>
-                              <span className="truncate">{option.label}</span>
-                            </button>
-                          ))}
-                        </div>
-                      )}
-                    </div>
-                  )}
-                  <button
-                    type="button"
-                    onClick={() => (batchMode ? exitBatch() : setBatchMode(true))}
-                    className={`h-8 px-3 rounded-full text-[13px] font-medium transition-colors ${
-                      batchMode
-                        ? (isDark ? 'bg-[#A8C7FA] text-[#041E49]' : 'bg-[#D3E3FD] text-[#041E49]')
-                        : (isDark ? 'text-[#C4C7C5] hover:bg-[#1E1F20]' : 'text-[#444746] hover:bg-[#F0F4F9]')
-                    }`}
-                  >
-                    {batchMode ? t.searchBatchDone : t.searchBatchManage}
-                  </button>
-                </div>
-              </div>
-
-              {/* 主体:左日期栏 + 右对话列表 */}
-              <div className="flex-1 min-h-0 flex gap-2">
-                {/* 日期栏(跟随「对话|已收纳」切换的数据集) */}
-                <div className="w-[150px] max-sm:w-[112px] shrink-0 overflow-y-auto custom-scrollbar pr-1 flex flex-col gap-0.5">
-                  <button
-                    type="button"
-                    onClick={() => handlePickDate('all')}
-                    className={`w-full h-9 px-3 shrink-0 flex items-center justify-between gap-2 rounded-[12px] text-[13px] transition-colors ${
-                      activeDate === 'all'
-                        ? (isDark ? 'bg-[#1E1F20] text-[#E3E3E3] font-medium' : 'bg-[#F0F4F9] text-[#1F1F1F] font-medium')
-                        : (isDark ? 'text-[#C4C7C5] hover:bg-[#1E1F20]' : 'text-[#444746] hover:bg-[#F0F4F9]')
-                    }`}
-                  >
-                    <span className="truncate">{t.searchDateAll}</span>
-                    <span className={`shrink-0 text-[12px] ${isDark ? 'text-[#9AA0A6]' : 'text-[#8A8F94]'}`}>{railTotal}</span>
-                  </button>
-                  {railGroups.map(g => {
-                    const active = g.key === activeDate;
-                    return (
-                      <button
-                        key={g.key}
-                        type="button"
-                        onClick={() => handlePickDate(g.key)}
-                        className={`w-full h-9 px-3 shrink-0 flex items-center justify-between gap-2 rounded-[12px] text-[13px] transition-colors ${
-                          active
-                            ? (isDark ? 'bg-[#1E1F20] text-[#E3E3E3] font-medium' : 'bg-[#F0F4F9] text-[#1F1F1F] font-medium')
-                            : (isDark ? 'text-[#C4C7C5] hover:bg-[#1E1F20]' : 'text-[#444746] hover:bg-[#F0F4F9]')
-                        }`}
-                      >
-                        <span className="truncate">{formatDateGroupLabel(g.key, language)}</span>
-                        <span className={`shrink-0 text-[12px] ${isDark ? 'text-[#9AA0A6]' : 'text-[#8A8F94]'}`}>{g.rows.length}</span>
-                      </button>
-                    );
-                  })}
-                </div>
-
-                {/* 对话列表:「对话|已收纳」共用日期分组渲染,仅行项目不同 */}
-                <div ref={listRef} className="relative flex-1 min-w-0 overflow-y-auto custom-scrollbar">
-                  {showArchived && archivedList.length === 0 && !searching ? (
-                    <div className={`px-4 py-10 text-center text-[14px] ${isDark ? 'text-[#9AA0A6]' : 'text-[#8A8F94]'}`}>
-                      {t.searchArchivedEmpty}
-                    </div>
-                  ) : searching ? (
-                    matchedGroups.length > 0 ? matchedGroups.map(g => (
-                      <div key={g.key} ref={el => { if (el) groupRefs.current[g.key] = el; }}>
-                        {renderDateHeader(g.key)}
-                        {g.rows.map(renderChatRow)}
-                      </div>
-                    )) : (
-                      <div className={`px-4 py-10 text-center text-[14px] ${isDark ? 'text-[#9AA0A6]' : 'text-[#8A8F94]'}`}>
-                        {t.searchNoResults}
-                      </div>
-                    )
-                  ) : (
-                    activeDate === 'all' ? dateGroups.map(g => (
-                      <div key={g.key}>
-                        {renderDateHeader(g.key)}
-                        {g.rows.map(renderChatRow)}
-                      </div>
-                    )) : (
-                      activeGroup && (
-                        <div>
-                          {renderDateHeader(activeGroup.key)}
-                          {activeGroup.rows.map(renderChatRow)}
-                        </div>
-                      )
-                    )
-                  )}
-                </div>
-              </div>
-
-              {/* 批量操作条:多选模式下吸附底部(在线=收纳/删除,已收纳=恢复/删除) */}
-              {batchMode && (
-                <div className={`absolute bottom-4 left-1/2 -translate-x-1/2 z-20 flex items-center gap-2 pl-4 pr-2 py-2 rounded-full shadow-xl border ${isDark ? 'bg-[#202124] border-white/10' : 'bg-white border-black/10'}`}>
-                  <span className={`text-[13px] whitespace-nowrap ${isDark ? 'text-[#C4C7C5]' : 'text-[#444746]'}`}>
-                    {t.searchSelectedCount(selectedIds.size)}
-                  </span>
-                  {batchDeleteConfirming ? (
-                    <span className="flex items-center gap-1" onClick={e => e.stopPropagation()}>
-                      <span className={`text-[13px] whitespace-nowrap ${isDark ? 'text-[#F28B82]' : 'text-[#C5221F]'}`}>{t.riDelQ}</span>
-                      <button
-                        type="button"
-                        title={t.riDelConfirm}
-                        disabled={selectedIds.size === 0}
-                        onClick={runBatchDelete}
-                        className={`w-8 h-8 rounded-full flex items-center justify-center transition-colors ${isDark ? 'text-[#F28B82] hover:bg-[#5c2b29]' : 'text-[#C5221F] hover:bg-[#FAD2CF]'}`}
-                      >
-                        <Check size={15} />
-                      </button>
-                      <button
-                        type="button"
-                        title={t.cpCancel}
-                        onClick={() => setBatchDeleteConfirming(false)}
-                        className={`w-8 h-8 rounded-full flex items-center justify-center transition-colors ${isDark ? 'text-[#C4C7C5] hover:bg-[#444746]' : 'text-[#5F6368] hover:bg-[#D3D7DB]'}`}
-                      >
-                        <X size={14} />
-                      </button>
-                    </span>
-                  ) : (
-                    <>
-                      <button
-                        type="button"
-                        disabled={selectedIds.size === 0}
-                        onClick={showArchived ? runBatchRestore : runBatchArchive}
-                        className={`h-8 px-3 rounded-full text-[13px] font-medium transition-colors disabled:opacity-40 ${isDark ? 'bg-[#A8C7FA] text-[#041E49]' : 'bg-[#D3E3FD] text-[#041E49]'}`}
-                      >
-                        {showArchived ? t.searchRestore : t.archiveSession}
-                      </button>
-                      <button
-                        type="button"
-                        disabled={selectedIds.size === 0}
-                        onClick={() => setBatchDeleteConfirming(true)}
-                        className={`h-8 px-3 rounded-full text-[13px] font-medium transition-colors disabled:opacity-40 ${isDark ? 'text-[#F28B82] hover:bg-[#5c2b29]' : 'text-[#C5221F] hover:bg-[#FAD2CF]'}`}
-                      >
-                        {t.cpDelete}
-                      </button>
-                    </>
-                  )}
-                </div>
-              )}
-
-              {/* 永久删除已收纳对话(二次确认,沿用原设置页同款弹窗) */}
-              {archivedDeleteConfirm && createPortal(
-                <ArchivedDeleteConfirmDialog
-                  theme={theme}
-                  t={t}
-                  onCancel={() => setArchivedDeleteConfirm(null)}
-                  onConfirm={() => {
-                    const id = archivedDeleteConfirm.id;
-                    setArchivedDeleteConfirm(null);
-                    if (id && onDelete) onDelete(id);
-                  }}
-                />,
-                document.body
-              )}
-
-            </div>
-          </div>
-        </div>
-      );
-    };
 
 
     // ==========================================
