+++ pinvou3-app/src/components/icons.jsx
@@ -100,8 +100,9 @@ const I = ({size=24,className='',children}) => (
     const Archive = (p) => <I {...p}><rect x="2" y="4" width="20" height="5" rx="1"/><path d="M4 9v10a1 1 0 0 0 1 1h14a1 1 0 0 0 1-1V9"/><path d="M10 13h4"/></I>;
     const PinIcon = (p) => <I {...p}><path d="M12 17v5"/><path d="M5 17h14"/><path d="M8 3h8"/><path d="M9 3l1 9-3 5h10l-3-5 1-9"/></I>;
     const PinOffIcon = (p) => <I {...p}><path d="M12 17v5"/><path d="M5 17h12"/><path d="M8 3h8"/><path d="M9 3l1 9-2 3"/><path d="M14 12l1-9"/><path d="M3 3l18 18"/></I>;
+    const PetPawIcon = (p) => <I size={18} {...p}><circle cx="11" cy="4" r="2"/><circle cx="18" cy="8" r="2"/><circle cx="4" cy="8" r="2"/><path d="M14.35 13.5a4 4 0 0 0-6.7 0c-.63 1.4-1.6 2.4-2.25 3.3-.72 1 .16 3.2 1.9 3.2 1.4 0 2.3-1 4.35-1s2.95 1 4.35 1c1.74 0 2.62-2.2 1.9-3.2-.65-.9-1.62-1.9-2.25-3.3z"/></I>;
     /* ==========================================
        i18n dictionary — 1:1 from 前端.html
        ========================================== */
 
-export { I, Plus, Edit2, Trash2, ClipboardList, BarChart2, Settings, Monitor, Smartphone, Brain, BrainCircuit, Clock, Sun, Moon, Zap, Package, RefreshCw, RotateCcw, Search, Lock, Play, Pause, PauseCircle, Upload, Lightbulb, Paperclip, Mic, Send, Store, Terminal, ChevronDown, IconGrid, IconList, ChevronRight, Copy, CheckCircle2, AlertTriangle, Menu, MoreHorizontal, Check, Filter, Database, Download, FolderPlus, Award, Feather, AppWindow, Maximize2, Minimize2, Radio, Palette, Briefcase, Sparkles, StopCircle, XCircle, Wrench, User, Layers, MessageSquare, X, ArrowLeft, FolderOpen, ExternalLink, Link, BookOpen, Code, FileText, Newspaper, ClipboardCheck, FileChartLine, Hexagon, Layout, Presentation, Mail, MessageCircle, Navigation, Video, Puzzle, LineChart, Building2, Cpu, Server, Globe, ChevronLeft, XIcon, CloudSun, TrendingUp, TrendingDown, GridIcon, TableIcon, PresentationIcon, ImageIcon, Archive, PinIcon, PinOffIcon };
+export { I, Plus, Edit2, Trash2, ClipboardList, BarChart2, Settings, Monitor, Smartphone, Brain, BrainCircuit, Clock, Sun, Moon, Zap, Package, RefreshCw, RotateCcw, Search, Lock, Play, Pause, PauseCircle, Upload, Lightbulb, Paperclip, Mic, Send, Store, Terminal, ChevronDown, IconGrid, IconList, ChevronRight, Copy, CheckCircle2, AlertTriangle, Menu, MoreHorizontal, Check, Filter, Database, Download, FolderPlus, Award, Feather, AppWindow, Maximize2, Minimize2, Radio, Palette, Briefcase, Sparkles, StopCircle, XCircle, Wrench, User, Layers, MessageSquare, X, ArrowLeft, FolderOpen, ExternalLink, Link, BookOpen, Code, FileText, Newspaper, ClipboardCheck, FileChartLine, Hexagon, Layout, Presentation, Mail, MessageCircle, Navigation, Video, Puzzle, LineChart, Building2, Cpu, Server, Globe, ChevronLeft, XIcon, CloudSun, TrendingUp, TrendingDown, GridIcon, TableIcon, PresentationIcon, ImageIcon, Archive, PinIcon, PinOffIcon, PetPawIcon };
